Add modules for Measurements section

+++ src/main/kotlin/ie/setu/domain/Measurement.kt
@@ -4,6 +4,6 @@ import org.joda.time.DateTime
 
 data class Measurement (var id: Int,
                         var bodyPart:String,
-                        var size: Int,
+                        var size: Double,
                         var measuredDate: DateTime,
                         var userId: Int)
