+++ src/main/kotlin/ie/setu/domain/db/Measurements.kt
@@ -9,7 +9,7 @@ import org.jetbrains.exposed.sql.Table
 object Measurements : Table("measurements") {
     val id = integer("id").autoIncrement().primaryKey()
     val bodyPart = varchar("body_part", 100)
-    val size = integer("size")
+    val size = double("size")
     val measuredDate = datetime("measured_date")
     val userId = integer("user_id").references(Users.id, onDelete = ReferenceOption.CASCADE)
 }
